Refactor PersonIndex into AnthologyIndex
