+++ bin/anthology/__init__.py
@@ -1,4 +1,4 @@
 from .anthology import Anthology
-from .people import PersonName, PersonIndex
+from .people import PersonName
 from .papers import Paper
 from .volumes import Volume
